Script component serialization - Ignored an error for opengl32 duplicate defines - Fixed loading classes with no namespace

+++ Locus/src/Locus/Utils/FileSystem.h
@@ -20,7 +20,7 @@ namespace Locus
 
 			std::streampos end = stream.tellg();
 			stream.seekg(0, std::ios::beg);
-			uint32_t size = end - stream.tellg();
+			uint32_t size = (uint32_t)(end - stream.tellg());
 
 			if (size == 0)
 			{
